use "," instead of " + " as separator
this is due to a potential bug in the html5 multifieldselector widget which can only recognize comma separated fields. this widget also does not seem to work with the default update select with multifield option. a bug report will be filed with the reporting team and escalated if this workaround is not backword compatible.